requirements.txt
torch==2.11.0
torchvision==0.26.0
torchaudio==2.11.0
numpy==2.2.6
matplotlib==3.10.9
pandas==2.3.3
scikit-learn==1.7.2
seaborn==0.13.2
pillow==12.2.0
opencv-python==4.13.0.92
tqdm==4.67.3
jupyter==1.1.1
ipykernel==7.2.0